+++ src/models/config_metrics_settings.rs
@@ -16,7 +16,7 @@ pub struct ConfigMetricsSettings {
     #[serde(rename = "Enable", skip_serializing_if = "Option::is_none")]
     pub enable: Option<bool>,
     #[serde(rename = "BlockProfileRate", skip_serializing_if = "Option::is_none")]
-    pub block_profile_rate: Option<i32>,
+    pub block_profile_rate: Option<i64>,
     #[serde(rename = "ListenAddress", skip_serializing_if = "Option::is_none")]
     pub listen_address: Option<String>,
 }
