+++ src/models/inline_response_200_16.rs
@@ -15,7 +15,7 @@
 pub struct InlineResponse20016 {
     /// The number of granular retention policies.
     #[serde(rename = "total_count", skip_serializing_if = "Option::is_none")]
-    pub total_count: Option<i32>,
+    pub total_count: Option<i64>,
 }
 
 impl InlineResponse20016 {
